% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_dates}
\alias{available_dates}
\title{List Available Dates for R and Bioconductor Releases}
\usage{
available_dates()
}
\value{
A character vector containing the available dates
}
\description{
List Available Dates for R and Bioconductor Releases
}
\examples{
available_dates()
}
\seealso{
Other available versions:
\code{\link[=available_df]{available_df()}},
\code{\link[=available_r]{available_r()}}
}
\concept{available versions}
