% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_r}
\alias{available_r}
\title{List Available R Versions from the rstats-on-nix Fork of Nixpkgs}
\usage{
available_r()
}
\value{
A character vector containing the available R versions.
}
\description{
List Available R Versions from the rstats-on-nix Fork of Nixpkgs
}
\examples{
available_r()
}
\seealso{
Other available versions:
\code{\link[=available_dates]{available_dates()}},
\code{\link[=available_df]{available_df()}}
}
\concept{available versions}
