% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{find_rev}
\alias{find_rev}
\title{find_rev Find the right Nix revision}
\usage{
find_rev(r_version)
}
\arguments{
\item{r_version}{Character. R version to look for, for example, "4.2.0"}
}
\value{
A character. The Nix revision to use
}
\description{
find_rev Find the right Nix revision
}
\examples{
find_rev("4.2.0")
}
