% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_flake.R
\name{flake_templates}
\alias{flake_templates}
\title{List available flake templates}
\usage{
flake_templates()
}
\value{
Character vector of available template names.
}
\description{
Lists all available flake templates that can be used with \code{flake()}.
}
\examples{
flake_templates()
}
