% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rix_helpers.R
\name{generate_rpkgs}
\alias{generate_rpkgs}
\title{generate_rpkgs Internal function that generates the string containing the
correct Nix expression to get R packages.}
\usage{
generate_rpkgs(rPackages, flag_rpkgs)
}
\arguments{
\item{rPackages}{Character, list of R packages to install.}

\item{flag_rpkgs}{Character, are there any R packages at all?}
}
\description{
generate_rpkgs Internal function that generates the string containing the
correct Nix expression to get R packages.
}
