% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{get_current}
\alias{get_current}
\title{get_current Get the current R version and packages}
\usage{
get_current()
}
\value{
A character. The commit hash of the latest nixpkgs-unstable revision
}
\description{
get_current Get the current R version and packages
}
