% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{get_date_from_version}
\alias{get_date_from_version}
\title{Get latest most recent date for given R version}
\usage{
get_date_from_version(r_version)
}
\value{
A character vector showing the version for a given date
}
\description{
Get latest most recent date for given R version
}
