% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{get_version_from_date}
\alias{get_version_from_date}
\title{Get latest R version for given date}
\usage{
get_version_from_date(date)
}
\value{
A character vector showing the version for a given date
}
\description{
Get latest R version for given date
}
