% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix_hash.R
\name{hash_url}
\alias{hash_url}
\title{Return the SRI hash of an URL with .tar.gz}
\usage{
hash_url(url)
}
\arguments{
\item{url}{String with URL ending with \code{.tar.gz}}
}
\description{
Return the SRI hash of an URL with .tar.gz
}
