% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{init}
\alias{init}
\title{Inititate and maintain an isolated, project-specific and pure R setup via Nix}
\usage{
init(project_path = ".", message_type = c("simple", "verbose"))
}
\arguments{
\item{project_path}{Character with folder path to the isolated nix-R project.
Defaults to \code{"."}, which is the current path in the working directory.
If the folder does not exist yet, it will be created.}

\item{message_type}{Character. Message type, defaults to \code{"simple"}, which
gives mimimal but sufficient feedback. Other values are currently
\code{"verbose"}, which gives more detailed diagnostics.}
}
\description{
Bootstraps an isolated project folder and adds sensible defaults to
achieve reproducibiltiy and isolated software dependencies with highest
run-time purity.
}
\details{
This general bootstrap helper does a couple of configurations that ensure
that the project runs as isolated and pure as possible by default.
Since a host RStudio session not launched via nix does not inherit
environmental variables from \code{.zshrc} (on macOS), we need to add the path of
the nixpkgs store to the \code{PATH} variable.
}
