% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_flake.R
\name{list_templates}
\alias{list_templates}
\title{List available flake templates}
\usage{
list_templates()
}
\value{
Character vector of available template names.
}
\description{
Lists all available flake templates that can be used with \code{init_flake()}.
}
\examples{
list_templates()
}
