% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{r_nix_revs}
\alias{r_nix_revs}
\title{r_nix_revs}
\format{
A data frame with 34 rows and 3 variables:
\describe{
\item{ version }{ R version }
\item{ revision }{ Nix revision containing this version of R }
\item{ date }{ Date when this revision was available }
}
}
\source{
Nix Package Version, by Marcelo Lazaroni
}
\usage{
r_nix_revs
}
\description{
Nix revisions with old R releases
}
\keyword{datasets}
