% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{read_renv_lock}
\alias{read_renv_lock}
\title{read_renv_lock}
\usage{
read_renv_lock(renv_lock_path = "renv.lock")
}
\arguments{
\item{renv_lock_path}{location of the renv.lock file, defaults to "renv.lock"}
}
\value{
the result of reading renv.lock with \link[jsonlite:read_json]{jsonlite::read_json}
}
\description{
Reads renv.lock if it exists and can be parsed as json.
}
