% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv_lock_pkgs}
\alias{renv_lock_pkgs}
\title{renv_lock_pkgs}
\usage{
renv_lock_pkgs(renv_lock_path = "renv.lock")
}
\arguments{
\item{renv_lock_path}{location of the renv.lock file from which to get the
list of packages, defaults to "renv.lock"}
}
\value{
a character vector of all the package names listed in the renv.lock
file proved in renv_lock_path.
}
\description{
gets the names of all the packages in an renv.lock file
}
\details{
This does NOT get the versions of those packages or handle transitive
dependencies in the renv.lock file not directly used in the project.
It simply returns a vector of all the package names.
}
\examples{

rix(r_pkgs = renv_lock_pkgs())

}
