% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv_lock_r_ver}
\alias{renv_lock_r_ver}
\title{renv_lock_r_ver}
\usage{
renv_lock_r_ver(renv_lock_path = "renv.lock")
}
\arguments{
\item{renv_lock_path}{location of the renv.lock file from which to get the
R version, defaults to "renv.lock"}
}
\value{
a length 1 chatacter vector with the version of R recorded in
renv.lock
}
\description{
renv_lock_r_ver
}
\examples{

rix(r_ver = renv_lock_r_ver())

}
