% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv_remote_pkgs}
\alias{renv_remote_pkgs}
\title{renv_remote_pkgs}
\usage{
renv_remote_pkgs(renv_lock_pkg_info, type = renv_lock_pkg_info$RemoteType)
}
\arguments{
\item{renv_lock_pkg_info}{a the list representation of a single package
entry from an renv.lock file.}

\item{type}{the type of remote package, defaults to the RemoteType of the
renv entry.
currently supported types: 'github' 'gitlab'
see \href{https://remotes.r-lib.org/}{remotes} for more.}
}
\value{
a list with three elements named:
"package_name", "repo_url", "commit"
}
\description{
Construct a list to be passed in a list to the git_pkgs argument of \link{rix}
The list returned contains the information necessary to have nix attempt to
build the package from an external repository.
}
\examples{
\dontrun{
renv_remote_pkgs(read_renv_lock()$Packages$renv)
}
}
