% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rix-package.R
\docType{package}
\name{rix-package}
\alias{rix}
\alias{rix-package}
\title{rix: Rix: Reproducible Environments With Nix}
\description{
Provides helper functions to create reproducible development environments using the Nix package manager.
}
\author{
\strong{Maintainer}: Bruno Rodrigues \email{brodriguesco@protonmail.com} (\href{https://orcid.org/0000-0002-3211-3689}{ORCID})

}
\keyword{internal}
