% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{sysdata}
\alias{sysdata}
\title{sysdata}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{ version }{ R version }
\item{ revision }{ Nix revision containing this version of R }
\item{ date }{ Date when this revision was available }
}
}
\source{
Nix Package Version, by Marcelo Lazaroni
}
\usage{
sysdata
}
\description{
Nix revisions with old R releases
}
\keyword{datasets}
