% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_nix_ga.R
\name{tar_nix_ga}
\alias{tar_nix_ga}
\title{Run a Targets Pipeline on GitHub Actions}
\usage{
tar_nix_ga()
}
\value{
Nothing, copies file to a directory.
}
\description{
Run a Targets Pipeline on GitHub Actions
}
\details{
This function puts a \code{.yaml} file inside the \verb{.github/workflows/}
folders on the root of your project. This workflow file will use the
projects \code{default.nix} file to generate the development environment on
GitHub Actions and will then run the projects \{targets\} pipeline. Make
sure to give read and write permissions to the GitHub Actions bot.
}
\examples{
\dontrun{
tar_nix_ga()
}
}
\seealso{
Other CI/CD:
\code{\link[=ga_cachix]{ga_cachix()}}
}
\concept{CI/CD}
