% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_dates}
\alias{available_dates}
\title{List available dates for R and Bioconductor releases}
\usage{
available_dates()
}
\value{
A character vector containing the available dates
}
\description{
List available dates for R and Bioconductor releases
}
\examples{
available_dates()
}
\seealso{
Other available versions: 
\code{\link{available_df}()},
\code{\link{available_r}()}
}
\concept{available versions}
