% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_dates}
\alias{available_dates}
\title{List available dates.}
\usage{
available_dates()
}
\value{
A character vector containing the available dates
}
\description{
List available dates.
}
\examples{
available_dates()
}
