% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_df}
\alias{available_df}
\title{Return data frame with R, Bioc versions and supported platforms}
\usage{
available_df()
}
\value{
A data frame
}
\description{
Return data frame with R, Bioc versions and supported platforms
}
\examples{
available_dates()
}
\seealso{
Other available versions: 
\code{\link{available_dates}()},
\code{\link{available_r}()}
}
\concept{available versions}
