% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R, R/included-datasets.R
\docType{data}
\name{available_r}
\alias{available_r}
\title{List available R versions from Nixpkgs}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{ version }{ R version }
\item{ revision }{ Nix revision containing this version of R }
\item{ date }{ Date when this revision was available }
}
}
\source{
Nix Package Version, by Marcelo Lazaroni
}
\usage{
available_r()

available_r
}
\value{
A character vector containing the available R versions.
}
\description{
Nix revisions with old R releases
}
\examples{
available_r()
}
\keyword{datasets}
