% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{available_r}
\alias{available_r}
\title{available_r List available R versions}
\usage{
available_r()
}
\value{
A character vector containing the available R versrions.
}
\description{
available_r List available R versions
}
\examples{
available_r()
}
