% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_r.R
\name{available_r}
\alias{available_r}
\title{List available R versions from Nixpkgs}
\usage{
available_r()
}
\value{
A character vector containing the available R versions.
}
\description{
List available R versions from Nixpkgs
}
\examples{
available_r()
}
