% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{nix_build}
\alias{nix_build}
\title{Invoke shell command \code{nix-build} from an R session}
\usage{
nix_build(
  nix_file = file.path("default.nix"),
  exec_mode = c("blocking", "non-blocking")
)
}
\arguments{
\item{nix_file}{File path to the \code{default.nix} file. The default is
a \code{default.nix} file in the current working directory of the current R
session.}

\item{exec_mode}{Either \code{"blocking"} (default) or \verb{"non-blocking}. This
will either block the R session while the \code{nix-build} shell command is
executed, or run \code{nix-build} in the background ("non-blocking").}
}
\value{
integer of the process ID (PID) of \code{nix-build} shell command
launched.
}
\description{
Invoke shell command \code{nix-build} from an R session
}
