% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{nix_build}
\alias{nix_build}
\title{Invoke shell command \code{nix-build} from an R session}
\usage{
nix_build(
  exec_mode = c("blocking", "non-blocking"),
  nix_file = file.path("default.nix")
)
}
\value{
invisible
}
\description{
Invoke shell command \code{nix-build} from an R session
}
