% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{nix_build}
\alias{nix_build}
\title{Invoke shell command \code{nix-build} from an R session}
\usage{
nix_build(project_path = ".", exec_mode = c("blocking", "non-blocking"))
}
\arguments{
\item{project_path}{Path to the folder where the \code{default.nix} file resides.
session.}

\item{exec_mode}{Either \code{"blocking"} (default) or \verb{"non-blocking}. This
will either block the R session while the \code{nix-build} shell command is
executed, or run \code{nix-build} in the background ("non-blocking").}
}
\value{
integer of the process ID (PID) of \code{nix-build} shell command
launched, if \code{nix_build()} call is assigned to an R object. Otherwise, it
will be returned as "invisible".
}
\description{
Invoke shell command \code{nix-build} from an R session
}
