% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv2nix}
\alias{renv2nix}
\title{renv2nix}
\usage{
renv2nix(
  renv_lock_path = "renv.lock",
  return_rix_call = FALSE,
  method = c("fast", "accurate"),
  ...
)
}
\arguments{
\item{renv_lock_path}{location of the renv.lock file, defaults to "renv.lock"}

\item{return_rix_call}{return the generated rix function call instead of
evaluating it this is for debugging purposes, defaults to FALSE}

\item{method}{the method of generating a nix environment from and renv.lock file.
"fast" is an inexact conversion which simply extracts the R version and a list of all the
packages in an renv.lock file and adds them to the \code{r_pkgs} argument of \link{rix}, unless they
are from external package repositories such as being installed directly from a github
repository in which case an attempt is made to handle them and pass them to the
\code{git_pkgs} argument of \link{rix}

Currently defaults to "fast", "accurate" is not yet implemented}

\item{...}{any other parameters to pass to \link{rix}}
}
\value{
nothing side effects only, unless \code{return_rix_call = TRUE} in which case an unevaluated
call to the \link{rix} function is returned
}
\description{
renv2nix
}
