% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv_lock_r_ver}
\alias{renv_lock_r_ver}
\title{renv_lock_r_ver}
\usage{
renv_lock_r_ver(renv_lock_path = "renv.lock", override_r_ver = NULL)
}
\arguments{
\item{renv_lock_path}{location of the renv.lock file from which to get the
R version, defaults to "renv.lock"}

\item{override_r_ver}{Character, override the R version defined in the
\code{renv.lock} file with another version. This is especially useful if
the \code{renv.lock} file lists a version of R not (yet) available through Nix.}
}
\value{
a length 1 character vector with the version of R recorded in
renv.lock
}
\description{
renv_lock_r_ver
}
\examples{
\dontrun{
rix(r_ver = renv_lock_r_ver())
}
}
