% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_helpers.R
\name{renv_remote_pkgs}
\alias{renv_remote_pkgs}
\title{renv_remote_pkgs}
\usage{
renv_remote_pkgs(renv_lock_remote_pkgs, host = NULL)
}
\arguments{
\item{renv_lock_remote_pkgs}{the list of package information from an renv.lock file.}

\item{host}{the host of remote package, defaults to NULL meaning the RemoteHost of the
renv entry will be used.
currently supported hosts: 'api.github.com' 'gitlab.com'
see \href{https://remotes.r-lib.org/}{remotes} for more.}
}
\value{
a list of lists with three elements named:
"package_name", "repo_url", "commit"
}
\description{
Construct a list to be passed the git_pkgs argument of \link{rix}
The list returned contains the information necessary to have nix attempt to
build the packages from their external repositories.
}
\examples{
\dontrun{
renv_remote_pkgs(read_renv_lock()$Packages)
}
}
