% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{rix}
\alias{rix}
\title{rix Build a reproducible development environment definition}
\usage{
rix(r_ver, pkgs, rstudio = FALSE, path = ".")
}
\arguments{
\item{r_ver}{Character. The required R version. Leave this argument empty if
you want to use the latest R version available. You can check which R
versions are available using \code{available_r}}

\item{pkgs}{Vector of characters. List the required packages for your
analysis here.}

\item{rstudio}{Logical, defaults to FALSE. If TRUE, RStudio gets installed in
this environment to enable interactive work.}

\item{path}{Character. Where to write \code{default.nix}.}
}
\value{
Nothing, this function only has the side-effect of writing a file
called "default.nix" in the working directory. This file contains the
instructions to build a reproducible environment using the Nix package
manager.
}
\description{
rix Build a reproducible development environment definition
}
\details{
This function will write a \code{default.nix} in the chosen path. Using
the Nix package manager, it is then possible to build a reproducible
development environment using the \code{nix-build} command in the path. This
environment will contain the chosen version of R and packages, and will not
interfere with any other installed version (via Nix or not) on your
machine. Every dependency, including both R package dependencies but also
system dependencies like compilers will get installed as well in that
environment. If you use RStudio for interactive work, then set the
\code{rstudio} parameter to \code{TRUE}. If you use another IDE (for example Emacs or
Visual Studio Code), you do not need to add it to the \code{default.nix} file,
you can simply use the version that is installed on your computer. To use
Visual Studio Code however, you need to install the \code{{languageserver}}
package, so don't forget to add it to the list of packages. Once you built
the environment using \code{nix-build}, you can drop into an interactive session
using \code{nix-shell}. See the "How-to Nix" vignette for more details.
}
