% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{rix}
\alias{rix}
\title{rix Generates a Nix expression that builds a reproducible development environment}
\usage{
rix(
  r_ver = "latest",
  r_pkgs = NULL,
  system_pkgs = NULL,
  git_pkgs = NULL,
  ide = "other",
  project_path = ".",
  overwrite = FALSE,
  print = FALSE,
  shell_hook = "R --vanilla"
)
}
\arguments{
\item{r_ver}{Character, defaults to "latest". The required R version, for example "4.0.0".
To use the latest version of R, use "latest", if you need the latest, bleeding edge version
of R and packages, then use "latest". You can check which R versions are available using \code{available_r}.
For reproducibility purposes, you can also provide a nixpkgs revision.}

\item{r_pkgs}{Vector of characters. List the required R packages for your
analysis here.}

\item{system_pkgs}{Vector of characters. List further software you wish to install that
are not R packages such as command line applications for example.}

\item{git_pkgs}{List. A list of packages to install from Git. See details for more information.}

\item{ide}{Character, defaults to "other". If you wish to use RStudio to work
interactively use "rstudio" or "code" for Visual Studio Code. For other editors,
use "other". This has been tested with RStudio, VS Code and Emacs. If other
editors don't work, please open an issue.}

\item{project_path}{Character, defaults to the current working directory. Where to write
\code{default.nix}, for example "/home/path/to/project".
The file will thus be written to the file "/home/path/to/project/default.nix".}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, overwrite the \code{default.nix}
file in the specified path.}

\item{print}{Logical, defaults to FALSE. If TRUE, print \code{default.nix} to console.}

\item{shell_hook}{Character, defaults to \code{"R --vanilla"}. Commands added to the shell_hook get
executed when the Nix shell starts (via \code{shellHook}). So by default, using \verb{nix-shell default.nix} will
start R. Set to NULL if you want bash to be started instead.}
}
\value{
Nothing, this function only has the side-effect of writing a file
called "default.nix" in the working directory. This file contains the
expression to build a reproducible environment using the Nix package
manager.
}
\description{
rix Generates a Nix expression that builds a reproducible development environment
}
\details{
This function will write a \code{default.nix} in the chosen path. Using
the Nix package manager, it is then possible to build a reproducible
development environment using the \code{nix-build} command in the path. This
environment will contain the chosen version of R and packages, and will not
interfere with any other installed version (via Nix or not) on your
machine. Every dependency, including both R package dependencies but also
system dependencies like compilers will get installed as well in that
environment. If you use RStudio for interactive work, then set the
\code{rstudio} parameter to \code{TRUE}. If you use another IDE (for example Emacs or
Visual Studio Code), you do not need to add it to the \code{default.nix} file,
you can simply use the version that is installed on your computer. Once you built
the environment using \code{nix-build}, you can drop into an interactive session
using \code{nix-shell}. See the "Building reproducible development environments with rix"
vignette for detailled instructions.
Packages to install from Github must be provided in a list of 4 elements:
"package_name", "repo_url", "branch_name" and "commit".
This argument can also be a list of lists of these 4 elements. It is also possible to install old versions
of packages by specifying a version. For example, to install the latest
version of \code{{AER}} but an old version of \code{{ggplot2}}, you could
write: \code{r_pkgs = c("AER", "ggplot2@2.2.1")}. Note
however that doing this could result in dependency hell, because an older
version of a package might need older versions of its dependencies, but other
packages might need more recent versions of the same dependencies. If instead you
want to use an environment as it would have looked at the time of \code{{ggplot2}}'s
version 2.2.1 release, then use the Nix revision closest to that date, by setting
\code{r_ver = "3.1.0"}, which was the version of R current at the time. This
ensures that Nix builds a completely coherent environment.
}
