% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cachix.R
\name{setup_cachix}
\alias{setup_cachix}
\title{setup_cachix Setup up cachix as a substituter}
\usage{
setup_cachix(nix_conf_path = "~/.config/nix")
}
\value{
Nothing, changes a file in the user's home directory.
}
\description{
setup_cachix Setup up cachix as a substituter
}
\details{
This function edits \verb{~/.config/nix/nix.conf} to add the
\code{rstats-on-nix} public cache as a substituter. The \code{rstats-on-nix} public
cache, hosted on Cachix, contains many prebuild binaries of R R packages
for x86_64 Linux and macOS (Intel architectures for packages released
before 2021 and Apple Silicon from 2021 onwards). This function
automatically performs a backup of \verb{~/.config/nix/nix.conf}.
}
\examples{
\dontrun{
setup_cachix()
}
}
