% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rev.R
\name{with_nix}
\alias{with_nix}
\title{Evaluate expression in R or default shell}
\usage{
with_nix(
  expr,
  program = c("R", "shell"),
  exec_mode = c("blocking", "non-blocking"),
  project_path = "."
)
}
\arguments{
\item{expr}{Single R function or call, shell command, or list of them.}

\item{program}{String stating where to evaluate the expression. Either \code{"R"},
the default, or \code{"shell"}. \code{where = "R"} will evaluate the expression via
\code{RScript} and \code{where = "shell"} will run in the standard shell.}

\item{exec_mode}{Either \code{"blocking"} (default) or \verb{"non-blocking}. This
will either block the R session while the \code{nix-build} shell command is
executed, or run \code{nix-build} in the background ("non-blocking").}

\item{project_path}{Path to the folder where the \code{default.nix} file resides.
The default is \code{"."}, which is the working directory in the current R
session.}
}
\description{
Evaluate expression in R or default shell
}
